+++ storage_drivers/ontap_san.go
@@ -185,6 +185,11 @@ func (d *OntapSANStorageDriver) Create(name string, sizeBytes uint64, opts map[s
 		return fmt.Errorf("Volume %s already exists.", name)
 	}
 
+	if sizeBytes < OntapMinimumVolumeSizeBytes {
+		return fmt.Errorf("Requested volume size (%d bytes) is too small.  The minimum volume size is %d bytes.",
+			sizeBytes, OntapMinimumVolumeSizeBytes)
+	}
+
 	// Get options with default fallback values
 	// see also: ontap_common.go#PopulateConfigurationDefaults
 	size := strconv.FormatUint(sizeBytes, 10)